move compress to common
