+++ pkg/vm/engine/tae/tables/indexwrapper/zmnode_test.go
@@ -24,6 +24,7 @@ import (
 	"github.com/matrixorigin/matrixone/pkg/fileservice"
 	"github.com/matrixorigin/matrixone/pkg/objectio"
 	"github.com/matrixorigin/matrixone/pkg/testutil"
+	"github.com/matrixorigin/matrixone/pkg/vm/engine/tae/common"
 	"github.com/matrixorigin/matrixone/pkg/vm/engine/tae/containers"
 	"github.com/matrixorigin/matrixone/pkg/vm/engine/tae/testutils"
 	"github.com/matrixorigin/matrixone/pkg/vm/mheap"
@@ -60,7 +61,7 @@ func TestBlockZoneMapIndex(t *testing.T) {
 	blocks, err := objectWriter.WriteEnd()
 	assert.Nil(t, err)
 	assert.Equal(t, 1, len(blocks))
-	cType := Plain
+	cType := common.Plain
 	typ := types.Type{Oid: types.T_int32}
 	pkColIdx := uint16(0)
 	interIdx := uint16(0)
